requirements.txt
APScheduler==3.6.3
async-timeout==4.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
idna==3.4
importlib-metadata==6.0.0
packaging==23.0
pymongo==4.1.1
python-dotenv==0.21.0
python-telegram-bot==13.11
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
rq==1.10.1
six==1.16.0
tornado==6.2
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
wrapt==1.14.1
zipp==3.11.0
